requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
supabase>=2.0.0
python-dotenv>=1.0.0
fastembed>=0.2.0